feat: implement MessageRepository for SQLite message storage and retrieval, and integrate with server logic
